requirements.txt
pymongo
pandas
networkx